requirements.txt
numpy
torch
pillow
torchvision
opencv-python
requests